Remove fields like DIFF and LVAL
The DIFF field is the same as dval-drbv, so remove DIFF and RDIF.
At the same time, correct the calculation of preferred_dir: when
the motor is moved whithin the backlash distance into the right direction,
we can just move to the new postion using backlash parameters.
Add code in TC 142 to test this.

Remove the fields LVAL, LDVL and LRVL and put them into the "private"
structure pmr->priv->last.