Simulator: Add rampUpAfterStart
Some controllers need one or two polls to return
moving after a movement has been started.
This is typically the case when the controller uses
multithreading, one thread for the motion control itself
and another one for the communication. If both threads are
not synchronized, the motion seams ready, but has not even started.
Note that this may happen only randomly.

Simulate this in the simulator.