Simulator: Simplify hw_motor_init()
Instead of passing a lot of paramaters into the function,
we use a struct, which is initiated with 0.
Values that need to be changed for this very parameter will
be set to a value != 0.

This allows more members to be added to the struct without the need
to change the source code for all drivers.